+++ src/docs/devguide/hugo/site/content/release-notes.md
@@ -1,5 +1,17 @@
 This page describes changes to the Giftcard Service Interface implemented across different releases of the interface.
 
+## v3.3.0
+
+Released 09 February 2018
+
+- Added Transfer functionality - allow funds to be transferred between gift cards.
+    - Added TransferRequest, TransferResponse, TransferConfirmation and TransferReversal.
+- Added Replace functionality - allows an old gift card to be replaced with a new gift card
+    - Added ReplaceRequest, ReplaceResponse, ReplaceConfirmation and ReplaceReversal.
+- Update ErrorDetail RequestTypes with new Replace and Transfer request operations
+- Updated giftcard-service-interface to latest service-base-interface (version 3.6.0)
+    - Updated GiftcardAmounts toString() method to include the additionalAmounts field
+ 
 ## v3.2.0
 
 Released 22 January 2018